tharindu
all changes

+++ demo/src/main/java/CanteenSystem/TransactionRepository.java
@@ -2,5 +2,5 @@ package CanteenSystem;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
 }
